… changelog.txt
… test/benchmarks/oldInspectVsNewInspect.jsx
… test/benchmarks/oldInspectVsNewInspect_Results.txt