Whew, this took a while …
but now I a quite happy with the results. 😎

# The all new `b.inspect()`

`b.inspect()` has been completely reworked to get rid of its flaws and improve it in general. I have been using this new version for a while now in the production of some scripts and it turns out to be extremely useful (and teaches you a lot about properties and methods 😉 ). The following improvements, fixes and changes have been made:

## Speed

The new implementation is a lot faster than the old one. For complex objects this means it can be more than 100x faster (that's 100 _times_, not 100 percent). This is achieved by avoiding a lot of recursion, as well as collecting all the results in a virtual string first, before it gets printed once in the end. I created a [benchmark file](https://github.com/trych/basil.js/blob/inspect/test/benchmarks/oldInspectVsNewInspect.jsx) that compares both functions, here are some results:

```
-----------------------------------------
Inspecting a Boolean (100 loops)
(old): 0.514 s
(new): 0.306 s
Speed up factor: 1.6797385620915
-----------------------------------------
-----------------------------------------
Inspecting Numbers (121 loops)
(old): 0.349 s
(new): 0.381 s
Speed up factor: 0.91601049868766
-----------------------------------------
-----------------------------------------
Inspecting a String (100 loops)
(old): 0.32 s
(new): 0.508 s
Speed up factor: 0.62992125984252
-----------------------------------------
-----------------------------------------
Inspecting a Custom Object @ Level 1 (10 loops)
(old): 0.262 s
(new): 0.036 s
Speed up factor: 7.27777777777778
-----------------------------------------
-----------------------------------------
Inspecting a Custom Object @ Level 2 (10 loops)
(old): 0.655 s
(new): 0.065 s
Speed up factor: 10.0769230769231
-----------------------------------------
-----------------------------------------
Inspecting an Array @ Level 1 (10 loops)
(old): 0.273 s
(new): 0.034 s
Speed up factor: 8.02941176470588
-----------------------------------------
-----------------------------------------
Inspecting an Array @ Level 2 (10 loops)
(old): 0.684 s
(new): 0.05 s
Speed up factor: 13.68
-----------------------------------------
-----------------------------------------
Inspecting a Color @ Level 1
(old): 0.042 s
(new): 0.006 s
Speed up factor: 7
-----------------------------------------
-----------------------------------------
Inspecting a Color @ Level 2
(old): 1.795 s
(new): 0.182 s
Speed up factor: 9.86263736263736
-----------------------------------------
-----------------------------------------
Inspecting a Color @ Level 3
(old): 3.479 s
(new): 1.57 s
Speed up factor: 19.1153846153846
-----------------------------------------
-----------------------------------------
Inspecting a Rectangle @ Level 1
(old): 0.115 s
(new): 0.063 s
Speed up factor: 1.82539682539683
-----------------------------------------
-----------------------------------------
Inspecting a Rectangle @ Level 2
(old): 0.706 s
(new): 0.234 s
Speed up factor: 3.01709401709402
-----------------------------------------
-----------------------------------------
Inspecting a Rectangle @ Level 3
(old): 4.573 s
(new): 2.511 s
Speed up factor: 19.542735042735
-----------------------------------------
-----------------------------------------
Inspecting App @ Level 1
(old): 74.464 s
(new): 0.243 s
Speed up factor: 306.43621399177
-----------------------------------------
-----------------------------------------
Inspecting $ @ Level 1
(old): 0.174 s
(new): 0.011 s
Speed up factor: 15.8181818181818
-----------------------------------------
```
The exact speed-up depends a lot on the type of object you are passing and on the `maxLevel` etc. The new function was only slower for inspecting simple strings and numbers and probably only because I chose to show their properties and methods during the tests. If you want to use the file to compare, you need to checkout [this commit](https://github.com/trych/basil.js/commit/9d54b16581defff3135d6bbfcec38aa25cbab262) that contains both the old and the new method next to each other.

## Presentation
The presentation of the results has been cleaned up a lot, so it should be much easier to read and understand what you are inspecting, especially for more complex objects. A tree style diagram is "drawn" to the console to present the results. I spent quite a while to try to achieve a look that is easily legible, I hope you guys like the result. The `PROPERTIES` and `METHODS` sections are highlighted for easy distinguishing and for each recursion level down you will see a new branch.

Example output of a simple nested array `["I", "love", "Basil", [true, false], 23, 5, b.color(255, 0, 0)]:

```js
[I, love, Basil, true,false, 23, 5, [object Color]] (Array)
|
|   PROPERTIES
|-- length: 7
|-- 0: I
|-- 1: love
|-- 2: Basil
|-- 3: Array (2)
|   |
|   |   PROPERTIES
|   |-- length: 2
|   |-- 0: true
|   |-- 1: false
|   
|-- 4: 23
|-- 5: 5
|-- 6: [object Color]
|   |
|   |   PROPERTIES
|   |-- colorValue: Array (3)
|   |-- eventListeners: [object EventListeners]
|   |-- events: [object Events]
|   |-- id: 215
|   |-- index: 10
|   |-- isValid: true
|   |-- isValid: true
|   |-- label: 
|   |-- model: PROCESS
|   |-- name: R=255 G=0 B=0
|   |-- parent: [object Document]
|   |-- properties: [object Object]
|   |-- space: RGB
|
|   METHODS
|-- concat()
|-- filter()
|-- join()
|-- map()
|-- pop()
|-- push()
|-- reverse()
|-- shift()
|-- slice()
|-- sort()
|-- splice()
|-- toLocaleString()
|-- toSource()
|-- toString()
|-- unshift()
```

The properties and methods are sorted alphabetically to find stuff easier. With complex objects the old method was almost unusable, the new method makes it much cleaner:

![inspectnewold](https://cloud.githubusercontent.com/assets/9803905/25683521/d68cc7be-305c-11e7-95d4-07d01d1e4ff5.jpg)

## Options

The new method has a `settings` parameter, which has parameters to control the result of `b.inspect()`. The parameters are (quoting from the new documentation):

- `showProps`: Show or hide properties. Default: `true`
- `showValues`: Show or hide values. Default: `true`
- `showMethods`: Show or hide methods. Default: `false`
- `maxLevel`: Chooses how many levels of properties should be inspected recursively. Default: `1`
- `propList`: Allows to pass an array of property names to show. If propList is not set all properties will be shown. Default: `[]` (no propList)

If no settings object is set, the default values will be used. The `propList` parameter is useful when I want to only observe a few specific properties of a complex object. For example, observe only the `geometricBounds` and `strokeWeight` values of a rect: `b.inspect(myRect, {maxLevel: 2, propList: ["geometricBounds", "strokeWeight"]};`.

Btw. I set the default for `showMethods` to `false` now, as in my tests I got the impression that it is a lot more confusing for beginners to have these methods around (as beginners probably don't know what methods even are). If they learn about methods at some point, they can always choose to show them. Is this okay for you guys?

## Reliability

The function is a lot more robust now and can handle all sorts of data types. The old version for example could not handle `null` or `undefined` values, which is something that could quite possibly be the value of a variable that is passed. Inspecting `null` or `undefined` works now. The new function also correctly shows the values of Enumerators, which might look like this:

```js
|-- strokeAlignment: CENTER_ALIGNMENT
```
while the old function would only state

```js
strokeAlignment: Enumerator
```

---
---

That's it. The code is quite complex, since a lot of exceptions need to be handled.

This closes #142. Please review. I think for reviewing you probably need to ~~load the branch~~ (well you actually don't need to load a branch anymore, you can just grab the bundled `basil.js` file from my branch – the perks of having basil in a single file 😉  ) and actually make use of the new `b.inspect()` function to test if it works for you.

@fabianmoronzirfas Could you also please carefully review the documentation of the improved method? I tried to follow the new jsDoc guidelines, I was not quite sure how to handle the settings object. I gave three examples to make it easy to understand how to use it with and without `settings` object and with the `propList`.