@ffd8 2017-05-03
Excited to try this! Ran into plenty of the annoying 'null' bug when inspecting something like the NYTimes API's in our recent class... will try to give feedback soon.