@fabianmoronzirfas 2017-05-04
To things I want to address.

1. In General: Would be great if you start working on branches directly in the basil.js repo instead of your fork. This makes checking out new branches for inspection/review easier. (Currently I keep a copy of your repo around as well)
2. About `inspect`: I would vote for making the `showMethods` by default `true`. This should not be hidden from the user. (I'm not sure if we already hat that discussion…)
3. **Awesome work.** The tree view looks already great. Do you know the [`tree`](http://mama.indstate.edu/users/ice/tree/) command (brew install tree). It gives you a listing of files in folders and their hierarchy. e.g.  

```txt
├── doc
├── extras
├── lib
├── node_modules
├── scripts
├── src
└── test
```
 
What I really like is that the last element is showing with the `└──` that it is the last. Might be cool for `b.inspect` as well.