@trych 2017-05-04
@fabianmoronzirfas 

> Would be great if you start working on branches directly in the basil.js repo instead of your fork.

I knew that day would come. 😉 Haha, no you're right, I will make the switch soon, just need to figure out how to transfer unfinished and experimental branches first.

> About inspect: I would vote for making the showMethods by default true. This should not be hidden from the user.

What do the others think? @b-g @ffd8 Personally I think it's less confusing for beginners, if it not shown by default, but if you all want it shown by default, no problem.

> Do you know the tree command (brew install tree).

Yes, I actually know it from some of your posts here on GitHub, that's also where the inspiration came from.

> What I really like is that the last element is showing with the └── that it is the last. Might be cool for b.inspect as well.

I totally agree, and I also tried this at first, the problem is, that the `Run in InDesign` package in Sublime (that I use exclusively) can only render ASCII characters at the moment. That means something like

```js
|-- 4: 23
|-- 5: 5
|-- 6: [object Color]
|   |
|   |   PROPERTIES
|   |-- colorValue: Array (3)
|   |-- eventListeners: [object EventListeners]
|   |-- events: [object Events]
|   |-- id: 215
|   |-- index: 10
|   |-- isValid: true
|   |-- isValid: true
|   |-- label: 
|   |-- model: PROCESS
|   |-- name: R=255 G=0 B=0
|   |-- parent: [object Document]
|   |-- properties: [object Object]
|   |-- space: RGB
|
|   METHODS
|-- concat()
|-- filter()
```

using these fancy `└──` characters would look something like this right now:

```js
??? 4: 23
??? 5: 5
??? 6: [object Color]
?   ?
?   ?   PROPERTIES
?   ??? colorValue: Array (3)
?   ??? eventListeners: [object EventListeners]
?   ??? events: [object Events]
?   ??? id: 215
?   ??? index: 10
?   ??? isValid: true
?   ??? isValid: true
?   ??? label: 
?   ??? model: PROCESS
?   ??? name: R=255 G=0 B=0
?   ??? parent: [object Document]
?   ??? properties: [object Object]
?   ??? space: RGB
?   METHODS
??? concat()
??? filter()
```

So that does not work obviously, so I tried to find the cleanest way using ASCII characters only instead. What I suggest we do: Merge it in its current (ASCII only) implementation for now, file an issue at the `Run in InDesign` repo to get the rendering of non-ASCII characters. And if this is done, then we could easily switch to proper `└──` characters. Should be a simple replacement of the strings, plus some small tweak for the last item of a branch, nothing dramatic. What do you think?