@trych 2017-05-09
@fabianmoronzirfas Moved the captions behind the examples tag.

Did anybody have any chance yet to test this out?