@fabianmoronzirfas 2017-05-09
I had a quick look but no extensive testing. Caught up in work around here. Sorry. But I think you did a lot of testing with it. It looks way better then the inspect before and I'm sure you tested it on any kind of object. So I would say go ahead and merge. @ffd8 @b-g What do you say?