@trych 2017-05-09
Yes, I did a lot of testing, I think it should work a lot better generally. It's a bit difficult to test every user scenario with this function, so I guess if there are still minor issues, they should become apparent it this function is used over time and people report back. So, yes, if nobody is generally against this function or its implementation, I would also vote to merge it.

@ff8d @b-g , do you agree?

And before I merge it, I still would like to know your opinions on the default of `showMethods`. Fabian wants it to show by default, I tend to not show it by default, what do the others think?