@trych 2017-05-09
Btw. is GitHub acting weird today or is this something strange on my end? None of the text formatting buttons work for me and I cannot edit my posts, because the edit button does not work for me, neither the emoji button. Did they change something recently?