@ffd8 2017-05-09
I was also having weird issues today with github - but only in an older version of safari = aaargh.

@trych finally sat down to test it (and agree about making it a branch in the main repo for easy testing- but now I've cloned you! so.. yeah)

I gave it a test from the standpoint of our recent in class project of NYTimes JSON API. The previous inspect failed whenever a null was reached, making it pretty useless. This new one is a major major improvement! But a few questions/suggestions:

- I'd drop the PROPERTIES word all together.. with that line gone, it will be so much cleaner to read. I'd keep that blank space as you have it from the parent (so removing PROPERTIES and thus, everything moves up one line).

- maxLevel is great! but it caused some weird issues while looking at a JSON that was quite simple.. it kept doing a 'reflect' that I don't understand. ie. here's an example JSON URL to try loading (just have to make a quick NYTimes API account = 1min)

```
  var apiKey = "insert_API_key";
  var apidate = "20170417";
  var nytURL = "http://api.nytimes.com/svc/community/v3/user-content/by-date.json?api-key="+apiKey+"&date="+apidate;
  var nytJSON = b.loadString(nytURL); 
  var nyt = b.JSON.decode(nytJSON); // decode json to object 
  b.inspect(nyt, {maxLevel:3});
```
it constantly produced this output in the inspect:
```
 PROPERTIES
|   |-- reflect: [object Reflection]
|   |   |
|   |   |   PROPERTIES
|   |   |-- description: 
|   |   |-- help: 
|   |   |-- methods: Array (9)
|   |   |-- name: Object
|   |   |-- properties: Array (5)
|   |   |-- sampleCode: null
|   |   |-- sampleFile: null
|   |   |-- staticMethods: Array (1)
|   |   |-- staticProperties: Array (3)
```

- showValues sounded like a great way to see the structure without the actual contents getting in the way, however when setting it to false, it ignored my setting of `maxLevels:3` – not sure if that's a bug or intentional, since it is ignoring the values... the question is if it could/should respect the request of values? or make an exception for the value being another object or array and going further? ignoring any value that isn't such a thing?

Just inspected a textframe = it's sooo beautiful! Really great work!